Put prompts at each step.

+++ arch-install-2.sh
@@ -4,59 +4,81 @@
 ## Date: June 11, 2014
 ## Arch Installer Part #2: System Configuration
 
-## Assumptions:
-#  - The base system has been installed using pacstrap.
-#  - We have chrooted into /mnt now
-
 ## Preamble
 # Define colors for 'tput'
 textrev=$(tput rev)
 textblue=$(tput setaf 1)
 textred=$(tput setaf 4)
 textreset=$(tput sgr0)
+note() {
+    echo ${textblue}$@${textreset}
+}
+header() {
+    echo ${textred}$@${textreset}
+}
+
 # Print script information
-echo ${rev}Arch Linux Installer [Script II - Configuration]${textreset}
-echo ${textred}Note: This script should be run from a chroot jail, where the base system has been installed using pacstrap.${textreset}
+header "Arch Linux Installer [Script II - Configuration]"
+note "This script is assumed to run from a chroot jail, where the base system has been installed using pacstrap."
+echo
 
 # The configuration begins here...
 
 ## Locale
-echo ${textblue}Setting Locale...${textreset}
+header "Step 1: Locale" 
+note "I will set the locale now."
+read -p "Do you wish to (c)ontinue/(S)kip/e(x)it? " -n 1 -r
 echo
+if [[ $REPLY =~ ^[Xx]$ ]] ; then
+	exit 0
+elif [[ $REPLY =~ ^[Yy]$ ]] ; then
+    # Set the correct Locale
+    mv /etc/locale.gen /etc/locale.gen.backup
+    echo en_US.UTF-8 UTF-8 > /etc/locale.gen
 
-# Set the correct Locale
-mv /etc/locale.gen /etc/locale.gen.backup
-echo en_US.UTF-8 UTF-8 > /etc/locale.gen
-# Generate the locale
-locale-gen
-echo LANG=en_US.UTF-8 > /etc/locale.conf
-export LANG=en_US.UTF-8
+    # Generate the locale
+    locale-gen
+    echo LANG=en_US.UTF-8 > /etc/locale.conf
+    export LANG=en_US.UTF-8
+fi
 echo
 
 ## Console font and keymap
 # Ignore this for now.
 
 ## Time zone
+header "Step 2: Time Zone"
 AIS_TIMEZONE=Asia/Calcutta
-echo ${textblue}Setting timezone to $AIS_TIMEZONE...${textreset}
-ln -s /usr/share/zoneinfo/$AIS_TIMEZONE /etc/localtime
+note "I will set the time zone to $AIS_TIMEZONE"
+read -p "Do you wish to (c)ontinue/(S)kip/e(x)it? " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Xx]$ ]] ; then
+	exit 0
+elif [[ $REPLY =~ ^[Yy]$ ]] ; then
+    note "Setting timezone to $AIS_TIMEZONE..."
+    ln -s /usr/share/zoneinfo/$AIS_TIMEZONE /etc/localtime
+fi
 echo
 
 ## Hardware clock
-echo ${textblue}Setting the Hardware clock to UTC...${textreset}
+header "Step 3: Hardware Clock" 
+note "Setting the Hardware clock to UTC..."
 hwclock --systohc --utc
 # Use the following for localtime (discouraged, used by Windows)
 # hwclock --systohv --localtime
 echo
 
 ## Kernel Modules
-# We'll go with the defaults for now.
+header "Step 4: Kernel Modules" 
+note "We'll go with the defaults for now. Skipping this step..."
+echo
 
 ## Hostname
+header "Step 5: Hostname" 
 AIS_HOST=dabba
-echo ${textblue}Setting hostname as ${AIS_HOST}${textreset}
+note "Setting hostname as ${AIS_HOST}..."
 echo $AIS_HOST > /etc/hostname
-echo ${textred}You should add this hostname to /etc/hosts.${textreset}
+note You should add this hostname to /etc/hosts.
 read -p "Do you want to edit /etc/hosts? (y/n) " -n 1 -r
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]
@@ -65,14 +87,10 @@ then
 fi
 echo
 
-## Network
-# We need to set up the network for the base system at this point. Maybe its better 
-# to shift this step to the top, so that no manual intervention is requitextred for this script.
-
 ## Initial Ramdisk Environment
-echo ${textblue}Creating an initial ramdisk environment...${textreset}
-echo You may want to edit /etc/mkinitcpio.conf. This is usually not requitextred.
-echo However, if you are installing this on a USB drive, put \"block\" right after \"udev\" in the HOOKS array.
+header "Step 6: Initial Ramdisk Environment"
+note "You may want to edit /etc/mkinitcpio.conf. This is usually not required."
+note "However, if you are installing this on a USB drive, you need to put \"block\" right after \"udev\" in the HOOKS array."
 read -p "Do you want to edit mkinitcpio.conf now? (y/n) " -n 1 -r
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]
@@ -83,36 +101,49 @@ fi
 echo
 
 ## Root password
-echo ${textblue}Setting the root password...${textreset}
-echo Please enter a password for the root account. 
+header "Step 7: Root Password" 
+note "Please enter a password for the root account."
 passwd
 echo
 
 ## Bootloader
-ai-install_bootloader () {
-    echo ${textblue}Installing the bootloader: syslinux...${textreset}
+ai_bootloader () {
     pacman -S syslinux
     syslinux-install_update -i -a -m
-    echo I will now open /boot/syslinux/syslinux.cfg for you.
-    echo You need to edit it to point to the correct root partition.
+
+    note "I will now open /boot/syslinux/syslinux.cfg for you."
+    note "You need to edit it to point to the correct root partition."
+    note "Remember to use UUID. You may use the following command to get the UUID of a partition: "
+    note "  blkid -s UUID -o value /dev/sdx1"
     read -p "Press any key to continue..  " 
     $EDITOR /boot/syslinux/syslinux.cfg
-    echo The bootloader has been installed. 
+    note The bootloader has been installed. 
     echo
-    echo ${textblue}This may be a good time to install something else too. Otherwise, please proceed to exit from chroot, unmount the root partition and reboot.${textreset}
-
 }
 
-## Check Network for the newly installed system.
-echo ${textblue}Checking if the network is up...${textreset}
-if ping -c 3 8.8.8.8; then
-    echo We have a working internet connection. Proceeding to install the bootloader.
-    echo
-    ai-install_bootloader
-else
-    echo ${textred}There seems to be a problem with the internet connection. Please configure the network properly.${textreset}
-    echo ${textblue}Once you have a working network connection, you may run ${textrev}ai-install_boatloader${textreset}${textblue} to complete the installation.${textreset}
+header "Step 8: Bootloader"
+note "I will install the bootloader Syslinux now." 
+read -p "Do you wish to (c)ontinue/(S)kip/e(x)it? " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Xx]$ ]] ; then
+	exit 0
+elif [[ $REPLY =~ ^[Yy]$ ]] ; then
+    header Checking network...
+    if ! ping -c 3 8.8.8.8 ; then
+        note "There seems to be some problem with your network. Please ensure that you have a working internet connection and try again."
+        exit 1
+    else
+        note "We have a working network. Proceeding to install the bootloader..." 
+        ai_bootloader
+    fi
 fi
+echo
+
+###########################
+header "Epilogue"
+note "This may be a good time to install something else too. For instance you may want to ensure that the network would work after rebooting. To make your life easier, install the packages wpa_supplicant and dialog for WiFi."
+note "After you are done, you may proceed to exit from chroot, unmount the newly installed root partition and reboot."
+note "To transform this bare bones installation to an complete and awesome linux system, run the Arch Installer Script III after you reboot." 
 
 ## Unmount and reboot
 #echo Exiting from chroot environment...
